Fixed fractured modifiers parsing and trade results
