+++ data/poe1/de/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Impliziter Modifikator",
+  "modDescriptionEnchantment": "Impliziter Modifikator (Verzauberung)",
+  "modDescriptionCorrupted": "Impliziter Modifikator (Verderbtheit)",
+  "modDescriptionEaterOfWorlds": "Impliziter Modifikator des Weltenfressers (#)",
+  "modDescriptionSearingExarch": "Impliziter Modifikator des Brennenden Exarchen (#)",
+  "modDescriptionPrefix": "Pr\u00E4fix-Modifikator \u0022#\u0022",
+  "modDescriptionSuffix": "Suffix-Modifikator \u0022#\u0022",
+  "modDescriptionCrafted": "Meisteranfertigung #",
+  "modDescriptionFractured": "Br\u00FCchig #",
+  "modDescriptionDesecrated": "Entweiht #",
+  "modDescriptionUnique": "Einzigartiger Modifikator",
+  "modDescriptionScourge": "Scourge-Modifikator",
+  "modDescriptionFoulborn": "F\u00E4ulnisgeborener einzigartiger Modifikator",
+  "modDescriptionTier": "(Level: #)",
+  "modDescriptionRank": "(Rang: #)",
+  "modDescriptionLevel": "(Stufe: #)",
+  "modDescriptionTierLevel": "(Level: #, Stufe: #)",
+  "modDescriptionRankLevel": "(Rang: #, Stufe: #)"
+}
