+++ data/poe1/en/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Implicit Modifier",
+  "modDescriptionEnchantment": "Enchantment Modifier",
+  "modDescriptionCorrupted": "Corruption Implicit Modifier",
+  "modDescriptionEaterOfWorlds": "Eater of Worlds Implicit Modifier (#)",
+  "modDescriptionSearingExarch": "Searing Exarch Implicit Modifier (#)",
+  "modDescriptionPrefix": "Prefix Modifier \u0022#\u0022",
+  "modDescriptionSuffix": "Suffix Modifier \u0022#\u0022",
+  "modDescriptionCrafted": "Master Crafted #",
+  "modDescriptionFractured": "Fractured #",
+  "modDescriptionDesecrated": "Desecrated #",
+  "modDescriptionUnique": "Unique Modifier",
+  "modDescriptionScourge": "Scourge Modifier",
+  "modDescriptionFoulborn": "Foulborn Unique Modifier",
+  "modDescriptionTier": "(Tier: #)",
+  "modDescriptionRank": "(Rank: #)",
+  "modDescriptionLevel": "(Lvl: #)",
+  "modDescriptionTierLevel": "(Tier: #, Lvl: #)",
+  "modDescriptionRankLevel": "(Rank: #, Lvl: #)"
+}
