+++ data/poe1/es/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Mod. impl\u00EDcito",
+  "modDescriptionEnchantment": "Mod. de encantamiento",
+  "modDescriptionCorrupted": "Mod. impl\u00EDcito de corrupci\u00F3n",
+  "modDescriptionEaterOfWorlds": "Mod. impl\u00EDcito del Devorador de mundos (#)",
+  "modDescriptionSearingExarch": "Mod. impl\u00EDcito del Exarca abrasador (#)",
+  "modDescriptionPrefix": "Mod. de prefijo \u0022#\u0022",
+  "modDescriptionSuffix": "Mod. de sufijo \u0022#\u0022",
+  "modDescriptionCrafted": "# fabricado por maestros",
+  "modDescriptionFractured": "# fracturado",
+  "modDescriptionDesecrated": "# profanado",
+  "modDescriptionUnique": "Mod. \u00FAnico",
+  "modDescriptionScourge": "Mod. de calamidad",
+  "modDescriptionFoulborn": "Mod. \u00FAnico impuro",
+  "modDescriptionTier": "(Grado: #)",
+  "modDescriptionRank": "(Rango: #)",
+  "modDescriptionLevel": "(Nivel: #)",
+  "modDescriptionTierLevel": "(Grado: #, Nivel: #)",
+  "modDescriptionRankLevel": "(Rango: #, Nivel: #)"
+}
