+++ data/poe1/fr/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Modificateur implicite",
+  "modDescriptionEnchantment": "Enchantement, Modificateur implicite",
+  "modDescriptionCorrupted": "Corruption, Modificateur implicite",
+  "modDescriptionEaterOfWorlds": "Implicite du Mangeur de mondes (#)",
+  "modDescriptionSearingExarch": "Implicite de l\u0027Exarque ardent (#)",
+  "modDescriptionPrefix": "Modificateur de pr\u00E9fixe : \u00AB # \u00BB",
+  "modDescriptionSuffix": "Modificateur de suffixe : \u00AB # \u00BB",
+  "modDescriptionCrafted": "Artisanat de ma\u00EEtre #",
+  "modDescriptionFractured": "Fissur\u00E9 #",
+  "modDescriptionDesecrated": "Profan\u00E9 #",
+  "modDescriptionUnique": "Modificateur d\u0027Objet unique",
+  "modDescriptionScourge": "Modificateur Souill\u00E9",
+  "modDescriptionFoulborn": "Modificateur d\u0027Objet unique malgendr\u00E9",
+  "modDescriptionTier": "(Palier : #)",
+  "modDescriptionRank": "(Rang : #)",
+  "modDescriptionLevel": "(Niv. : #)",
+  "modDescriptionTierLevel": "(Palier : #, Niv. : #)",
+  "modDescriptionRankLevel": "(Rang : #, Niv. : #)"
+}
