+++ data/poe1/ja/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\u6697\u9ED9\u30E2\u30C3\u30C9",
+  "modDescriptionEnchantment": "\u30A8\u30F3\u30C1\u30E3\u30F3\u30C8\u30E2\u30C3\u30C9",
+  "modDescriptionCorrupted": "\u30B3\u30E9\u30D7\u30C8\u6697\u9ED9\u30E2\u30C3\u30C9",
+  "modDescriptionEaterOfWorlds": "\u30A4\u30FC\u30BF\u30FC\u30FB\u30AA\u30D6\u30FB\u30EF\u30FC\u30EB\u30BA \u6697\u9ED9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionSearingExarch": "\u30B7\u30A2\u30EA\u30F3\u30B0\u30FB\u30A8\u30B0\u30B6\u30FC\u30AF \u6697\u9ED9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionPrefix": "\u30D7\u30EC\u30D5\u30A3\u30C3\u30AF\u30B9\u30E2\u30C3\u30C9\u300C#\u300D",
+  "modDescriptionSuffix": "\u30B5\u30D5\u30A3\u30C3\u30AF\u30B9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionCrafted": "\u30DE\u30B9\u30BF\u30FC\u30AF\u30E9\u30D5\u30C8 #",
+  "modDescriptionFractured": "\u30D5\u30E9\u30AF\u30C1\u30E3\u30FC #",
+  "modDescriptionDesecrated": "\u5192\u6D9C #",
+  "modDescriptionUnique": "\u30E6\u30CB\u30FC\u30AF\u30E2\u30C3\u30C9",
+  "modDescriptionScourge": "\u30B9\u30AB\u30FC\u30B8\u30E2\u30C3\u30C9",
+  "modDescriptionFoulborn": "\u30D5\u30A1\u30A6\u30EB\u30DC\u30FC\u30F3\u30E6\u30CB\u30FC\u30AF\u30E2\u30C3\u30C9",
+  "modDescriptionTier": "(\u30C6\u30A3\u30A2: #)",
+  "modDescriptionRank": "(\u30E9\u30F3\u30AF: #)",
+  "modDescriptionLevel": "(\u30EC\u30D9\u30EB: #)",
+  "modDescriptionTierLevel": "(\u30C6\u30A3\u30A2: #, \u30EC\u30D9\u30EB: #)",
+  "modDescriptionRankLevel": "(\u30E9\u30F3\u30AF: #, \u30EC\u30D9\u30EB: #)"
+}
