+++ data/poe1/ko/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\uACE0\uC815 \uC18D\uC131 \uBD80\uC5EC",
+  "modDescriptionEnchantment": "\uC18D\uC131 \uBD80\uC5EC \uC778\uCC48\uD2B8",
+  "modDescriptionCorrupted": "\uD0C0\uB77D \uACE0\uC815 \uC18D\uC131 \uBD80\uC5EC",
+  "modDescriptionEaterOfWorlds": "\uC138\uACC4 \uD3EC\uC2DD\uC790 \uACE0\uC815 \uC18D\uC131 (#)",
+  "modDescriptionSearingExarch": "\uC791\uC5F4\uC758 \uCD1D\uC8FC\uAD50 \uACE0\uC815 \uC18D\uC131 (#)",
+  "modDescriptionPrefix": "\uC811\uB450\uC5B4 \uC18D\uC131 \uBD80\uC5EC \u0022#\u0022",
+  "modDescriptionSuffix": "\uC811\uBBF8\uC5B4 \uC18D\uC131 \uBD80\uC5EC \u0022#\u0022",
+  "modDescriptionCrafted": "\uB300\uAC00\uC758 \uC81C\uC791 #",
+  "modDescriptionFractured": "\uBD84\uC5F4\uB41C #",
+  "modDescriptionDesecrated": "\uD6FC\uC190\uB41C #",
+  "modDescriptionUnique": "\uACE0\uC720 \uC18D\uC131 \uBD80\uC5EC",
+  "modDescriptionScourge": "\uC2A4\uCEEC\uC9C0 \uC18D\uC131 \uBD80\uC5EC",
+  "modDescriptionFoulborn": "\uC0BF\uB41C \uACE0\uC720 \uC18D\uC131",
+  "modDescriptionTier": "(\uB4F1\uAE09: #)",
+  "modDescriptionRank": "(\uB2E8\uACC4: #)",
+  "modDescriptionLevel": "(\uB808\uBCA8: #)",
+  "modDescriptionTierLevel": "(\uB4F1\uAE09: #, \uB808\uBCA8: #)",
+  "modDescriptionRankLevel": "(\uB2E8\uACC4: #, \uB808\uBCA8: #)"
+}
