+++ data/poe1/pt/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Mod Impl\u00EDcito",
+  "modDescriptionEnchantment": "Encantamento de Mod",
+  "modDescriptionCorrupted": "Mod Impl\u00EDcito Corrompido",
+  "modDescriptionEaterOfWorlds": "Modificador Impl\u00EDcito do Devorador de Mundos #",
+  "modDescriptionSearingExarch": "Modificador Impl\u00EDcito do Exarca Cauterizado #",
+  "modDescriptionPrefix": "Mod Prefixo \u0022#\u0022",
+  "modDescriptionSuffix": "Mod Sufixo \u0022#\u0022",
+  "modDescriptionCrafted": "Criado por Mestre: #",
+  "modDescriptionFractured": "Rompido: #",
+  "modDescriptionDesecrated": "Profanado: #",
+  "modDescriptionUnique": "Mod \u00DAnico",
+  "modDescriptionScourge": "Mod Castigado",
+  "modDescriptionFoulborn": "Modificador \u00DAnico Natimpuro",
+  "modDescriptionTier": "(Tier: #)",
+  "modDescriptionRank": "(Rank: #)",
+  "modDescriptionLevel": "(Lvl: #)",
+  "modDescriptionTierLevel": "(Tier: #, Lvl: #)",
+  "modDescriptionRankLevel": "(Rank: #, Lvl: #)"
+}
