+++ data/poe1/th/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\u0E21\u0E47\u0E2D\u0E14\u0E41\u0E1D\u0E07",
+  "modDescriptionEnchantment": "\u0E21\u0E47\u0E2D\u0E14\u0E40\u0E2D\u0E19\u0E41\u0E0A\u0E19\u0E17\u0E4C",
+  "modDescriptionCorrupted": "\u0E21\u0E47\u0E2D\u0E14\u0E41\u0E1D\u0E07\u0E21\u0E25\u0E17\u0E34\u0E19",
+  "modDescriptionEaterOfWorlds": "\u0E21\u0E47\u0E2D\u0E14\u0E41\u0E1D\u0E07\u0E1C\u0E39\u0E49\u0E01\u0E25\u0E37\u0E19\u0E01\u0E34\u0E19\u0E42\u0E25\u0E01\u0E32 (#)",
+  "modDescriptionSearingExarch": "\u0E21\u0E47\u0E2D\u0E14\u0E41\u0E1D\u0E07\u0E2D\u0E38\u0E1B\u0E23\u0E32\u0E0A\u0E2D\u0E32\u0E1A\u0E40\u0E1E\u0E25\u0E34\u0E07 (#)",
+  "modDescriptionPrefix": "\u0E21\u0E47\u0E2D\u0E14\u0E1E\u0E23\u0E35\u0E1F\u0E34\u0E01\u0E0B\u0E4C \u0022#\u0022",
+  "modDescriptionSuffix": "\u0E21\u0E47\u0E2D\u0E14\u0E0B\u0E31\u0E1F\u0E1F\u0E34\u0E01\u0E0B\u0E4C \u0022#\u0022",
+  "modDescriptionCrafted": "# \u0E17\u0E35\u0E48\u0E16\u0E39\u0E01\u0E04\u0E23\u0E32\u0E1F\u0E17\u0E4C",
+  "modDescriptionFractured": "# \u0E41\u0E15\u0E01\u0E23\u0E49\u0E32\u0E27",
+  "modDescriptionDesecrated": "# \u0E2A\u0E32\u0E21\u0E32\u0E19\u0E22\u0E4C",
+  "modDescriptionUnique": "\u0E21\u0E47\u0E2D\u0E14\u0E22\u0E39\u0E19\u0E34\u0E04",
+  "modDescriptionScourge": "\u0E21\u0E47\u0E2D\u0E14\u0E21\u0E34\u0E15\u0E34\u0E2D\u0E2A\u0E39\u0E23",
+  "modDescriptionFoulborn": "\u0E21\u0E47\u0E2D\u0E14\u0E44\u0E2D\u0E40\u0E17\u0E21\u0E22\u0E39\u0E19\u0E34\u0E04\u0E0A\u0E31\u0E48\u0E27\u0E0A\u0E49\u0E32",
+  "modDescriptionTier": "(\u0E2D\u0E31\u0E19\u0E14\u0E31\u0E1A: #)",
+  "modDescriptionRank": "(\u0E23\u0E30\u0E14\u0E31\u0E1A: #)",
+  "modDescriptionLevel": "(\u0E40\u0E25\u0E40\u0E27\u0E25: #)",
+  "modDescriptionTierLevel": "(\u0E2D\u0E31\u0E19\u0E14\u0E31\u0E1A: #, \u0E40\u0E25\u0E40\u0E27\u0E25: #)",
+  "modDescriptionRankLevel": "(\u0E23\u0E30\u0E14\u0E31\u0E1A: #, \u0E40\u0E25\u0E40\u0E27\u0E25: #)"
+}
