+++ data/poe1/zh/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\u56FA\u5B9A\u8A5E\u7DB4",
+  "modDescriptionEnchantment": "\u9644\u9B54\u56FA\u5B9A\u8A5E\u7DB4",
+  "modDescriptionCorrupted": "\u5DF2\u6C59\u67D3\u56FA\u5B9A\u8A5E\u7DB4",
+  "modDescriptionEaterOfWorlds": "\u541E\u566C\u5929\u5730\u56FA\u5B9A\u8A5E\u7DB4 (#)",
+  "modDescriptionSearingExarch": "\u707C\u70D9\u7E3D\u7763\u56FA\u5B9A\u8A5E\u7DB4 (#)",
+  "modDescriptionPrefix": "\u524D\u7DB4 \u0022#\u0022",
+  "modDescriptionSuffix": "\u5F8C\u7DB4 \u0022#\u0022",
+  "modDescriptionCrafted": "\u5DF2\u5927\u5E2B\u5DE5\u85DD #",
+  "modDescriptionFractured": "\u5DF2\u7834\u88C2 #",
+  "modDescriptionDesecrated": "\u5DF2\u893B\u7006 #",
+  "modDescriptionUnique": "\u50B3\u5947\u8A5E\u7DB4",
+  "modDescriptionScourge": "\u707D\u9B58\u8A5E\u7DB4",
+  "modDescriptionFoulborn": "Foulborn Unique Modifier",
+  "modDescriptionTier": "(\u968E\u5C64\uFF1A#)",
+  "modDescriptionRank": "(\u968E\u7D1A\uFF1A#)",
+  "modDescriptionLevel": "(\u7B49\u7D1A: #)",
+  "modDescriptionTierLevel": "(\u968E\u7D1A: #, \u7B49\u7D1A: #)",
+  "modDescriptionRankLevel": "(\u968E\u7D1A: #, \u7B49\u7D1A: #)"
+}
