+++ data/poe2/de/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Impliziter Modifikator",
+  "modDescriptionEnchantment": "Verst\u00E4rkung",
+  "modDescriptionCorrupted": "Verderbnis-Verst\u00E4rkung",
+  "modDescriptionEaterOfWorlds": "Impliziter Modifikator des Weltenfressers (#)",
+  "modDescriptionSearingExarch": "Impliziter Modifikator des Brennenden Exarchen (#)",
+  "modDescriptionPrefix": "Pr\u00E4fix-Modifikator \u0022#\u0022",
+  "modDescriptionSuffix": "Suffix-Modifikator \u0022#\u0022",
+  "modDescriptionCrafted": "Angefertigt #",
+  "modDescriptionFractured": "Br\u00FCchig #",
+  "modDescriptionDesecrated": "Entweiht #",
+  "modDescriptionUnique": "Einzigartiger Modifikator",
+  "modDescriptionVaalUnique": "Einzigartiger Vaal-Modifikator",
+  "modDescriptionScourge": "Scourge-Modifikator",
+  "modDescriptionTier": "(Level: #)",
+  "modDescriptionRank": "(Rang: #)",
+  "modDescriptionLevel": "(Stufe: #)",
+  "modDescriptionTierLevel": "(Level: #, Stufe: #)",
+  "modDescriptionRankLevel": "(Rang: #, Stufe: #)"
+}
