+++ data/poe2/en/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Implicit Modifier",
+  "modDescriptionEnchantment": "Enhancement",
+  "modDescriptionCorrupted": "Corruption Enhancement",
+  "modDescriptionEaterOfWorlds": "Eater of Worlds Implicit Modifier (#)",
+  "modDescriptionSearingExarch": "Searing Exarch Implicit Modifier (#)",
+  "modDescriptionPrefix": "Prefix Modifier \u0022#\u0022",
+  "modDescriptionSuffix": "Suffix Modifier \u0022#\u0022",
+  "modDescriptionCrafted": "Crafted #",
+  "modDescriptionFractured": "Fractured #",
+  "modDescriptionDesecrated": "Desecrated #",
+  "modDescriptionUnique": "Unique Modifier",
+  "modDescriptionVaalUnique": "Vaal Unique Modifier",
+  "modDescriptionScourge": "Scourge Modifier",
+  "modDescriptionTier": "(Tier: #)",
+  "modDescriptionRank": "(Rank: #)",
+  "modDescriptionLevel": "(Lvl: #)",
+  "modDescriptionTierLevel": "(Tier: #, Lvl: #)",
+  "modDescriptionRankLevel": "(Rank: #, Lvl: #)"
+}
