+++ data/poe2/es/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Mod. impl\u00EDcito",
+  "modDescriptionEnchantment": "Mejor\u00EDa",
+  "modDescriptionCorrupted": "Mejor\u00EDa de corrupci\u00F3n",
+  "modDescriptionEaterOfWorlds": "Mod. impl\u00EDcito del Devorador de mundos (#)",
+  "modDescriptionSearingExarch": "Mod. impl\u00EDcito del Exarca abrasador (#)",
+  "modDescriptionPrefix": "Mod. de prefijo \u0022#\u0022",
+  "modDescriptionSuffix": "Mod. de sufijo \u0022#\u0022",
+  "modDescriptionCrafted": "# de fabricaci\u00F3n",
+  "modDescriptionFractured": "# fracturado",
+  "modDescriptionDesecrated": "# profanado",
+  "modDescriptionUnique": "Mod. \u00FAnico",
+  "modDescriptionVaalUnique": "Modificador \u00FAnico vaal",
+  "modDescriptionScourge": "Mod. de calamidad",
+  "modDescriptionTier": "(Grado: #)",
+  "modDescriptionRank": "(Rango: #)",
+  "modDescriptionLevel": "(Nivel: #)",
+  "modDescriptionTierLevel": "(Grado: #, Nivel: #)",
+  "modDescriptionRankLevel": "(Rango: #, Nivel: #)"
+}
