+++ data/poe2/fr/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Modificateur Implicite",
+  "modDescriptionEnchantment": "Am\u00E9lioration",
+  "modDescriptionCorrupted": "Am\u00E9lioration de Corruption",
+  "modDescriptionEaterOfWorlds": "Implicite du Mangeur de mondes (#)",
+  "modDescriptionSearingExarch": "Implicite de l\u0027Exarque ardent (#)",
+  "modDescriptionPrefix": "Pr\u00E9fixe : \u00AB # \u00BB",
+  "modDescriptionSuffix": "Suffixe : \u00AB # \u00BB",
+  "modDescriptionCrafted": "Artisanat #",
+  "modDescriptionFractured": "Fissur\u00E9 #",
+  "modDescriptionDesecrated": "Profan\u00E9 #",
+  "modDescriptionUnique": "Modificateur d\u0027Objet unique",
+  "modDescriptionVaalUnique": "Modificateur vaal unique",
+  "modDescriptionScourge": "Modificateur Souill\u00E9",
+  "modDescriptionTier": "(Palier : #)",
+  "modDescriptionRank": "(Rang : #)",
+  "modDescriptionLevel": "(Niv. : #)",
+  "modDescriptionTierLevel": "(Palier : #, Niv. : #)",
+  "modDescriptionRankLevel": "(Rang : #, Niv. : #)"
+}
