+++ data/poe2/ja/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\u6697\u9ED9\u30E2\u30C3\u30C9",
+  "modDescriptionEnchantment": "\u30A8\u30F3\u30CF\u30F3\u30B9",
+  "modDescriptionCorrupted": "\u30B3\u30E9\u30D7\u30C8\u30A8\u30F3\u30CF\u30F3\u30B9",
+  "modDescriptionEaterOfWorlds": "\u4E16\u754C\u3092\u55B0\u3089\u3046\u8005 \u6697\u9ED9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionSearingExarch": "\u707C\u71B1\u306E\u4EE3\u884C\u8005 \u6697\u9ED9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionPrefix": "\u30D7\u30EC\u30D5\u30A3\u30C3\u30AF\u30B9\u30E2\u30C3\u30C9\u300C#\u300D",
+  "modDescriptionSuffix": "\u30B5\u30D5\u30A3\u30C3\u30AF\u30B9\u30E2\u30C3\u30C9 \u300C#\u300D",
+  "modDescriptionCrafted": "\u30AF\u30E9\u30D5\u30C8 #",
+  "modDescriptionFractured": "\u30D5\u30E9\u30AF\u30C1\u30E3\u30FC #",
+  "modDescriptionDesecrated": "\u5192\u6D9C #",
+  "modDescriptionUnique": "\u30E6\u30CB\u30FC\u30AF\u30E2\u30C3\u30C9",
+  "modDescriptionVaalUnique": "\u30F4\u30A1\u30FC\u30EB\u30E6\u30CB\u30FC\u30AF\u30E2\u30C3\u30C9",
+  "modDescriptionScourge": "\u30B9\u30AB\u30FC\u30B8\u30E2\u30C3\u30C9",
+  "modDescriptionTier": "(\u30C6\u30A3\u30A2: #)",
+  "modDescriptionRank": "(\u30E9\u30F3\u30AF: #)",
+  "modDescriptionLevel": "(\u30EC\u30D9\u30EB: #)",
+  "modDescriptionTierLevel": "(\u30C6\u30A3\u30A2: #, \u30EC\u30D9\u30EB: #)",
+  "modDescriptionRankLevel": "(\u30E9\u30F3\u30AF: #, \u30EC\u30D9\u30EB: #)"
+}
