+++ data/poe2/pt/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "Modificador Impl\u00EDcito",
+  "modDescriptionEnchantment": "Melhoria",
+  "modDescriptionCorrupted": "Melhoria de Corrup\u00E7\u00E3o",
+  "modDescriptionEaterOfWorlds": "Modificador Impl\u00EDcito do Devorador de Mundos (#)",
+  "modDescriptionSearingExarch": "Modificador Impl\u00EDcito do Exarca Calcinante (#)",
+  "modDescriptionPrefix": "Modificador de Prefixo \u0022#\u0022",
+  "modDescriptionSuffix": "Modificador de Sufixo \u0022#\u0022",
+  "modDescriptionCrafted": "Criou #",
+  "modDescriptionFractured": "Rompido: #",
+  "modDescriptionDesecrated": "Profanado: #",
+  "modDescriptionUnique": "Modificador \u00DAnico",
+  "modDescriptionVaalUnique": "Modificador \u00DAnico Vaal",
+  "modDescriptionScourge": "Modificador do Flagelo",
+  "modDescriptionTier": "(N\u00EDvel: #)",
+  "modDescriptionRank": "(Ranque: #)",
+  "modDescriptionLevel": "(Nv.: #)",
+  "modDescriptionTierLevel": "(Categoria: #, Nv.: #)",
+  "modDescriptionRankLevel": "(Ranque: #, Nv.: #)"
+}
