+++ data/poe2/ru/texts.json
@@ -1,0 +1,20 @@
+{
+  "modDescriptionImplicit": "\u0421\u043E\u0431\u0441\u0442\u0432\u0435\u043D\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E",
+  "modDescriptionEnchantment": "\u0423\u043B\u0443\u0447\u0448\u0435\u043D\u0438\u0435",
+  "modDescriptionCorrupted": "\u0423\u043B\u0443\u0447\u0448\u0435\u043D\u0438\u0435 \u043E\u0442 \u043E\u0441\u043A\u0432\u0435\u0440\u043D\u0435\u043D\u0438\u044F",
+  "modDescriptionEaterOfWorlds": "\u0421\u043E\u0431\u0441\u0442\u0432\u0435\u043D\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E \u041F\u043E\u0436\u0438\u0440\u0430\u0442\u0435\u043B\u044F \u043C\u0438\u0440\u043E\u0432 (#)",
+  "modDescriptionSearingExarch": "\u0421\u043E\u0431\u0441\u0442\u0432\u0435\u043D\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E \u041F\u043B\u0430\u043C\u0435\u043D\u043D\u043E\u0433\u043E \u044D\u043A\u0437\u0430\u0440\u0445\u0430 (#)",
+  "modDescriptionPrefix": "\u041F\u0440\u0435\u0444\u0438\u043A\u0441 \u0022#\u0022",
+  "modDescriptionSuffix": "\u0421\u0443\u0444\u0444\u0438\u043A\u0441 \u0022#\u0022",
+  "modDescriptionCrafted": "\u0420\u0435\u043C\u0435\u0441\u043B\u0435\u043D\u043D\u043E\u0435 #",
+  "modDescriptionFractured": "\u0420\u0430\u0441\u043A\u043E\u043B\u043E\u0442\u043E\u0435 #",
+  "modDescriptionDesecrated": "\u041E\u0447\u0435\u0440\u043D\u0451\u043D\u043D\u043E\u0435 #",
+  "modDescriptionUnique": "\u0423\u043D\u0438\u043A\u0430\u043B\u044C\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E",
+  "modDescriptionVaalUnique": "\u0423\u043D\u0438\u043A\u0430\u043B\u044C\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E \u0432\u0430\u0430\u043B",
+  "modDescriptionScourge": "\u041F\u0440\u0435\u043E\u0431\u0440\u0430\u0436\u0451\u043D\u043D\u043E\u0435 \u0441\u0432\u043E\u0439\u0441\u0442\u0432\u043E",
+  "modDescriptionTier": "(\u0423\u0440\u043E\u0432\u0435\u043D\u044C: #)",
+  "modDescriptionRank": "(\u0420\u0430\u043D\u0433: #)",
+  "modDescriptionLevel": "(\u0423\u0440.: #)",
+  "modDescriptionTierLevel": "(\u0420\u0430\u043D\u0433: #, \u0423\u0440.: #)",
+  "modDescriptionRankLevel": "(\u0420\u0430\u043D\u0433: #, \u0423\u0440.: #)"
+}
