+++ src/Sidekick.Data/Texts/DataText.cs
@@ -1,0 +1,24 @@
+﻿namespace Sidekick.Data.Texts;
+
+public class DataText
+{
+    public string? ModDescriptionImplicit { get; set; }
+    public string? ModDescriptionEnchantment { get; set; }
+    public string? ModDescriptionCorrupted { get; set; }
+    public string? ModDescriptionEaterOfWorlds { get; set; }
+    public string? ModDescriptionSearingExarch { get; set; }
+    public string? ModDescriptionPrefix { get; set; }
+    public string? ModDescriptionSuffix { get; set; }
+    public string? ModDescriptionCrafted { get; set; }
+    public string? ModDescriptionFractured { get; set; }
+    public string? ModDescriptionDesecrated { get; set; }
+    public string? ModDescriptionUnique { get; set; }
+    public string? ModDescriptionVaalUnique { get; set; }
+    public string? ModDescriptionScourge { get; set; }
+    public string? ModDescriptionFoulborn { get; set; }
+    public string? ModDescriptionTier { get; set; }
+    public string? ModDescriptionRank { get; set; }
+    public string? ModDescriptionLevel { get; set; }
+    public string? ModDescriptionTierLevel { get; set; }
+    public string? ModDescriptionRankLevel { get; set; }
+}
